double

+++ utils.h
@@ -13,11 +13,11 @@ void printAlgorithmMenu(int startPoint, int endPoint);
 void clearStream();
 int inputNumber(int min, int max);
 string inputFileName();
-vector<vector <int>> loadMatrix(ifstream& fin);
-void saveMatrix(const vector<vector <int>>& matrix, ofstream& fout);
-void printMatrix(const vector<vector <int>>& matrix);
-vector<vector <int>> createMatrix();
+vector<vector <double>> loadMatrix(ifstream& fin);
+void saveMatrix(const vector<vector <double>>& matrix, ofstream& fout);
+void printMatrix(const vector<vector <double>>& matrix);
+vector<vector <double>> createMatrix();
 int selectStartPoint(int size);
 int selectEndPoint(int size);
-int findShortcut(const vector<vector <int>>& matrix);
+int findShortcut(const vector<vector <double>>& matrix);
